+++ src/assets/resources/job-try-again-button.js
@@ -31,7 +31,7 @@ CraftliltPlugin.JobTryAgainButton = Garnish.Base.extend({
       location.reload();
     }).catch(exception => {
       Craft.cp.displayError(
-          Craft.t('app', 'Can\'t submit review, unexpected issue occurred'));
+          Craft.t('app', 'Can\'t mark reviewed, unexpected issue occurred'));
       this.$spinner.remove();
       $(this).removeClass('disabled');
     });