+++ src/controllers/translation/PostTranslationPublishController.php
@@ -26,26 +26,30 @@ class PostTranslationPublishController extends AbstractJobController
     public function actionInvoke(): Response
     {
         $request = Craft::$app->getRequest();
-        $translationId = $request->getBodyParam('translationId');
+        $translationIds = $request->getBodyParam('translationIds');
 
-        if (empty($translationId)) {
+        if (empty($translationIds)) {
             return (new Response())->setStatusCode(400);
         }
 
-        $translation = Craftliltplugin::getInstance()->translationRepository->findOneById((int)$translationId);
+        $translations = TranslationRecord::findAll(
+            ['id' => $translationIds]
+        );
 
-        if ($translation === null) {
+        if (empty($translations)) {
             return (new Response())->setStatusCode(404);
         }
 
-        Craftliltplugin::getInstance()->publishDraftsHandler->__invoke(
-            $translation->translatedDraftId,
-            $translation->targetSiteId
-        );
+        foreach ($translations as $translation) {
+            Craftliltplugin::getInstance()->publishDraftsHandler->__invoke(
+                $translation->translatedDraftId,
+                $translation->targetSiteId
+            );
+        }
 
         $updated = TranslationRecord::updateAll(
             ['status' => TranslationRecord::STATUS_PUBLISHED],
-            ['id' => $translation->id]
+            ['id' => $translationIds]
         );
 
         if ($updated !== 1) {
@@ -53,14 +57,16 @@ class PostTranslationPublishController extends AbstractJobController
         }
 
         if ($updated) {
-            Craftliltplugin::getInstance()->jobLogsRepository->create(
-                $translation->jobId,
-                Craft::$app->getUser()->getId(),
-                sprintf('Translation (id: %d) published', $translation->id)
-            );
+            foreach ($translations as $translation) {
+                Craftliltplugin::getInstance()->jobLogsRepository->create(
+                    $translation->jobId,
+                    Craft::$app->getUser()->getId(),
+                    sprintf('Translation (id: %d) published', $translation->id)
+                );
+            }
 
             Craftliltplugin::getInstance()->refreshJobStatusHandler->__invoke(
-                $translation->jobId
+                $translations[0]->jobId
             );
         }