+++ tests/integration/controllers/job/GetJobEditFormControllerCest.php
@@ -109,6 +109,17 @@ class GetJobEditFormControllerCest
         $controller = $this->getController();
         $controller->setView(new ViewWrapper());
 
+        $I->expectSettingsGetRequest(
+            '/api/v1.0/settings',
+            'SECURE_API_KEY_FOR_LILT_CONNECTOR',
+            [
+                'project_prefix' => 'this-is-connector-project-prefix',
+                'project_name_template' => 'this-is-connector-project-name-template',
+                'lilt_translation_workflow' => SettingsResponse::LILT_TRANSLATION_WORKFLOW_INSTANT,
+            ],
+            200
+        );
+
         $response = $controller->actionInvoke((string) $job->id);
 
         $expected = $this->getExpected($job);